move [legacy_]query_result.rs to response module
