+++ examples/cqlsh-rs.rs
@@ -5,8 +5,8 @@ use rustyline::{CompletionType, Config, Context, Editor};
 use rustyline_derive::{Helper, Highlighter, Hinter, Validator};
 use scylla::connection::Compression;
 use scylla::frame::response::result::Row;
+use scylla::response::query_result::IntoRowsResultError;
 use scylla::session::Session;
-use scylla::transport::query_result::IntoRowsResultError;
 use scylla::QueryResult;
 use scylla::SessionBuilder;
 use std::env;
