+++ scylla/src/response/mod.rs
@@ -1,3 +1,5 @@
+pub mod legacy_query_result;
+pub mod query_result;
 mod request_response;
 
 pub(crate) use request_response::{
