+++ scylla/src/transport/mod.rs
@@ -1,5 +1,2 @@
-pub mod legacy_query_result;
-pub mod query_result;
-
 pub use crate::frame::Authenticator;
 pub use scylla_cql::frame::request::query::{PagingState, PagingStateResponse};
